Rearrange register modules
